moved accounts unit tests to accountsLegacyTests.js, made a new accounts.js file in the test/unit/core location to be less integrated unit tests for the accounts module.
